requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
html5lib==1.0.1
importlib-metadata==1.6.1
Markdown==3.2.2
Pillow==7.1.2
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
zipp==3.1.0
